v0.3.0版本: 1. 添加单元测试; 2. 添加自动化测试; 3. 补充部分注释; 4. 优化代码并修复 bug;

+++ agtms-parent/agtms-mongodb/src/main/java/net/saisimon/agtms/mongodb/service/TemplateMongodbService.java
@@ -28,6 +28,9 @@ public class TemplateMongodbService implements TemplateService, MongodbOrder {
 	
 	@Override
 	public Template saveOrUpdate(Template entity) {
+		if (entity == null) {
+			return entity;
+		}
 		if (entity.getId() == null) {
 			Long id = sequenceService.nextId(templateMongodbRepository.getCollectionName());
 			entity.setId(id);
@@ -37,6 +40,9 @@ public class TemplateMongodbService implements TemplateService, MongodbOrder {
 
 	@Override
 	public boolean dropTable(Template template) {
+		if (template == null) {
+			return false;
+		}
 		mongoTemplate.dropCollection(TemplateUtils.getTableName(template));
 		return true;
 	}
